# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface CreateRoomRequest {
  hostName: string;
  teamName: string;
}

export interface JoinRoomRequest {
  name: string;
  teamName: string;
}

export interface JoinRoomResponse {
  userId: number;
  roomId: string;
  name: string;
  teamName: string;
  isHost: boolean;
}

export interface StartAuctionRequest {
  userId: number;
}

export interface PlaceBidRequest {
  userId: number;
  amount: number;
}

export interface BidResponse {
  success: boolean;
  currentBid: number;
  highestBidderId: number;
  timeRemaining: number;
}

export interface RoomUser {
  id: number;
  name: string;
  teamName: string;
  balance: number;
  squadCount: number;
  isHost: boolean;
}

export type RoomResponseStatus =
  (typeof RoomResponseStatus)[keyof typeof RoomResponseStatus];

export const RoomResponseStatus = {
  waiting: "waiting",
  active: "active",
  finished: "finished",
} as const;

export interface RoomResponse {
  roomId: string;
  status: RoomResponseStatus;
  users: RoomUser[];
}

export type PlayerCategory =
  (typeof PlayerCategory)[keyof typeof PlayerCategory];

export const PlayerCategory = {
  uncapped: "uncapped",
  capped: "capped",
  goat: "goat",
} as const;

export interface Player {
  id: number;
  name: string;
  role: string;
  country: string;
  category: PlayerCategory;
  basePrice: number;
  powerScore: number;
}

export type AuctionStateStatus =
  (typeof AuctionStateStatus)[keyof typeof AuctionStateStatus];

export const AuctionStateStatus = {
  waiting: "waiting",
  active: "active",
  finished: "finished",
} as const;

export interface AuctionState {
  roomId: string;
  status: AuctionStateStatus;
  currentPlayer: Player | null;
  currentBid: number | null;
  highestBidderId: number | null;
  highestBidderName: string | null;
  timeRemaining: number | null;
  playersAuctioned: number;
  totalPlayers: number;
  users: RoomUser[];
}

export interface SquadPlayer {
  id: number;
  name: string;
  role: string;
  country: string;
  category: string;
  basePrice: number;
  powerScore: number;
  pricePaid: number;
}

export interface SquadResponse {
  userId: number;
  name: string;
  teamName: string;
  balance: number;
  squadCount: number;
  squad: SquadPlayer[];
  totalPowerScore: number;
}

export interface ResultEntry {
  rank: number;
  userId: number;
  name: string;
  teamName: string;
  totalPowerScore: number;
  squadCount: number;
  balance: number;
  squad: SquadPlayer[];
}

export interface ResultsResponse {
  roomId: string;
  results: ResultEntry[];
}