# packages/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new room
 */
export const CreateRoomBody = zod.object({
  hostName: zod.string(),
  teamName: zod.string(),
});

/**
 * @summary Get room details
 */
export const GetRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  roomId: zod.string(),
  status: zod.enum(["waiting", "active", "finished"]),
  users: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      teamName: zod.string(),
      balance: zod.number(),
      squadCount: zod.number(),
      isHost: zod.boolean(),
    }),
  ),
});

/**
 * @summary Join a room
 */
export const JoinRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const JoinRoomBody = zod.object({
  name: zod.string(),
  teamName: zod.string(),
});

export const JoinRoomResponse = zod.object({
  userId: zod.number(),
  roomId: zod.string(),
  name: zod.string(),
  teamName: zod.string(),
  isHost: zod.boolean(),
});

/**
 * @summary Start the auction (host only)
 */
export const StartAuctionParams = zod.object({
  roomId: zod.coerce.string(),
});

export const StartAuctionBody = zod.object({
  userId: zod.number(),
});

export const StartAuctionResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Place a bid
 */
export const PlaceBidParams = zod.object({
  roomId: zod.coerce.string(),
});

export const PlaceBidBody = zod.object({
  userId: zod.number(),
  amount: zod.number(),
});

export const PlaceBidResponse = zod.object({
  success: zod.boolean(),
  currentBid: zod.number(),
  highestBidderId: zod.number(),
  timeRemaining: zod.number(),
});

/**
 * @summary Get current auction state
 */
export const GetAuctionStateParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetAuctionStateResponse = zod.object({
  roomId: zod.string(),
  status: zod.enum(["waiting", "active", "finished"]),
  currentPlayer: zod
    .object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      country: zod.string(),
      category: zod.enum(["uncapped", "capped", "goat"]),
      basePrice: zod.number(),
      powerScore: zod.number(),
    })
    .nullable(),
  currentBid: zod.number().nullable(),
  highestBidderId: zod.number().nullable(),
  highestBidderName: zod.string().nullable(),
  timeRemaining: zod.number().nullable(),
  playersAuctioned: zod.number(),
  totalPlayers: zod.number(),
  users: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      teamName: zod.string(),
      balance: zod.number(),
      squadCount: zod.number(),
      isHost: zod.boolean(),
    }),
  ),
});

/**
 * @summary Get a user's squad
 */
export const GetUserSquadParams = zod.object({
  roomId: zod.coerce.string(),
  userId: zod.coerce.number(),
});

export const GetUserSquadResponse = zod.object({
  userId: zod.number(),
  name: zod.string(),
  teamName: zod.string(),
  balance: zod.number(),
  squadCount: zod.number(),
  squad: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      country: zod.string(),
      category: zod.string(),
      basePrice: zod.number(),
      powerScore: zod.number(),
      pricePaid: zod.number(),
    }),
  ),
  totalPowerScore: zod.number(),
});

/**
 * @summary Get auction results
 */
export const GetResultsParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetResultsResponse = zod.object({
  roomId: zod.string(),
  results: zod.array(
    zod.object({
      rank: zod.number(),
      userId: zod.number(),
      name: zod.string(),
      teamName: zod.string(),
      totalPowerScore: zod.number(),
      squadCount: zod.number(),
      balance: zod.number(),
      squad: zod.array(
        zod.object({
          id: zod.number(),
          name: zod.string(),
          role: zod.string(),
          country: zod.string(),
          category: zod.string(),
          basePrice: zod.number(),
          powerScore: zod.number(),
          pricePaid: zod.number(),
        }),
      ),
    }),
  ),
});