# packages/api-zod/src/generated/types/auctionState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */
import type { AuctionStateStatus } from "./auctionStateStatus";
import type { Player } from "./player";
import type { RoomUser } from "./roomUser";

export interface AuctionState {
  roomId: string;
  status: AuctionStateStatus;
  currentPlayer: Player | null;
  currentBid: number | null;
  highestBidderId: number | null;
  highestBidderName: string | null;
  timeRemaining: number | null;
  playersAuctioned: number;
  totalPlayers: number;
  users: RoomUser[];
}