# packages/api-zod/src/generated/types/bidResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */

export interface BidResponse {
  success: boolean;
  currentBid: number;
  highestBidderId: number;
  timeRemaining: number;
}