# packages/api-zod/src/generated/types/joinRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinRoomRequest {
  name: string;
  teamName: string;
}