# packages/api-zod/src/generated/types/joinRoomResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinRoomResponse {
  userId: number;
  roomId: string;
  name: string;
  teamName: string;
  isHost: boolean;
}