# packages/api-zod/src/generated/types/resultEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */
import type { SquadPlayer } from "./squadPlayer";

export interface ResultEntry {
  rank: number;
  userId: number;
  name: string;
  teamName: string;
  totalPowerScore: number;
  squadCount: number;
  balance: number;
  squad: SquadPlayer[];
}