# packages/api-zod/src/generated/types/resultsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */
import type { ResultEntry } from "./resultEntry";

export interface ResultsResponse {
  roomId: string;
  results: ResultEntry[];
}