# packages/api-zod/src/generated/types/roomUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */

export interface RoomUser {
  id: number;
  name: string;
  teamName: string;
  balance: number;
  squadCount: number;
  isHost: boolean;
}