# packages/api-zod/src/generated/types/squadPlayer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IPL Fantasy Auction API
 * OpenAPI spec version: 0.1.0
 */

export interface SquadPlayer {
  id: number;
  name: string;
  role: string;
  country: string;
  category: string;
  basePrice: number;
  powerScore: number;
  pricePaid: number;
}